Cs {Hmisc}	R Documentation

Character strings from unquoted names

Description

Cs makes a vector of character strings from a list of valid R names. .q is similar but also makes uses of names of arguments.

Usage

Cs(...)
.q(...)

Arguments

...
any number of names separated by commas. For .q any names of arguments will be used.

Value

character string vector. For .q there will be a names attribute to the vector if any names appeared in ....

See Also

sys.frame, deparse

Examples

Cs(a,cat,dog)
# subset.data.frame <- dataframe[,Cs(age,sex,race,bloodpressure,height)]
.q(a, b, c, 'this and that')
.q(dog=a, giraffe=b, cat=c)

[Package Hmisc version 4.8-0 Index]