ebpcomp {Hmisc}	R Documentation

ebpcomp

Description

Computation of Coordinates of Extended Box Plots Elements

Usage

ebpcomp(x, qref = c(0.5, 0.25, 0.75), probs = c(0.05, 0.125, 0.25, 0.375))

Arguments

x
a numeric variable

qref
quantiles for major corners

probs
quantiles for minor corners

Details

For an extended box plots computes all the elements needed for plotting it. This is typically used when adding to a ggplot2 plot.

Value

list with elements segments, lines, points, points2

Author(s)

Frank Harrell

Examples

ebpcomp(1:1000)

[Package Hmisc version 5.1-2 Index]