format.pval {Hmisc}	R Documentation

Format P Values

Description

format.pval is intended for formatting p-values.

Usage

format.pval(x, pv=x, digits = max(1, .Options$digits - 2),
            eps = .Machine$double.eps, na.form = "NA", ...)

Arguments

pv
a numeric vector.

x
argument for method compliance.

digits
how many significant digits are to be used.

eps
a numerical tolerance: see Details.

na.form
character representation of NAs.

...
arguments passed to format in the format.pval function body.

Details

format.pval is mainly an auxiliary function for print.summary.lm etc., and does separate formatting for fixed, floating point and very small values; those less than eps are formatted as “< [eps]” (where “[eps]” stands for format(eps, digits)).

Value

A character vector.

Note

This is the base format.pval function with the ablitiy to pass the nsmall argument to format

Examples

format.pval(c(runif(5), pi^-100, NA))
format.pval(c(0.1, 0.0001, 1e-27))
format.pval(c(0.1, 1e-27), nsmall=3)

[Package Hmisc version 4.8-0 Index]