getZip {Hmisc}	R Documentation

Open a Zip File From a URL

Description

Allows downloading and reading of a zip file containing one file

Usage

getZip(url, password=NULL)

Arguments

url
either a path to a local file or a valid URL.

password
required to decode password-protected zip files

Details

Allows downloading and reading of zip file containing one file. The file may be password protected. If a password is needed then one will be requested unless given.

Note: to make password-protected zip file z.zip, do zip -e z myfile

Value

Returns a file O/I pipe.

Author(s)

Frank E. Harrell

See Also

pipe

Examples

## Not run: 
read.csv(getZip('http://test.com/z.zip'))

## End(Not run)

[Package Hmisc version 4.8-0 Index]