getabd {Hmisc}	R Documentation

getabd

Description

Data from The Analysis of Biological Data by Shitlock and Schluter

Usage

getabd(name = "", lowernames = FALSE, allow = "_")

Arguments

name
name of dataset to fetch. Omit to get a data table listing all available datasets.

lowernames
set to TRUE to change variable names to lower case

allow
set to NULL to convert underscores in variable names to periods

Details

Fetches csv files for exercises in the book

Value

data frame with attributes label and url

Author(s)

Frank Harrell

[Package Hmisc version 5.1-2 Index]