hidingTOC {Hmisc}	R Documentation

Moving and Hiding Table of Contents

Description

Moving and hiding table of contents for Rmd HTML documents

Usage

hidingTOC(
  buttonLabel = "Contents",
  levels = 3,
  tocSide = c("right", "left"),
  buttonSide = c("right", "left"),
  posCollapse = c("margin", "top", "bottom"),
  hidden = FALSE
)

Arguments

buttonLabel
the text on the button that hides and unhides the table of contents. Defaults to Contents.

levels
the max depth of the table of contents that it is desired to have control over the display of. (defaults to 3)

tocSide
which side of the page should the table of contents be placed on. Can be either 'right' or 'left'. Defaults to 'right'

buttonSide
which side of the page should the button that hides the TOC be placed on. Can be either 'right' or 'left'. Defaults to 'right'

posCollapse
if 'margin' then display the depth select buttons vertically along the side of the page choosen by buttonSide. If 'top' then display the depth select buttons horizontally under the button that hides the TOC. Defaults to 'margin'. 'bottom' is currently unimplemented.

hidden
Logical should the table of contents be hidden at page load Defaults to FALSE

Details

hidingTOC creates a table of contents in a Rmd document that can be hidden at the press of a button. It also generate buttons that allow the hiding or unhiding of the diffrent level depths of the table of contents. For quarto-related methods see https://stackoverflow.com/questions/73999697.

Value

a HTML formated text string to be inserted into an markdown document

Author(s)

Thomas Dupont

Examples

## Not run: 
hidingTOC()

## End(Not run)

[Package Hmisc version 4.8-0 Index]