htmltabv {Hmisc}	R Documentation

htmltabc

Description

Simple HTML Table of Verbatim Output

Usage

htmltabv(..., cols = 2, propts = list(quote = FALSE))

Arguments

...
objects to print(). The arguments must be named with the labels you want to print before the verbatim print().

cols
number of columns in the html table

propts
an option list of arguments to pass to the print() methods; default is to not quote character strings

Details

Uses capture.output to capture as character strings the results of running print() on each element of .... If an element of ... has length of 1 and is a blank string, nothing is printed for that cell other than its name (not in verbatim).

Value

character string of html

Author(s)

Frank Harrell

[Package Hmisc version 5.1-2 Index]