latestFile {Hmisc}	R Documentation

latestFile

Description

Find File With Latest Modification Time

Usage

latestFile(pattern, path = ".", verbose = TRUE)

Arguments

pattern
a regular expression; see base::list.files()

path
full path, defaulting to current working directory

verbose
set to FALSE to not report on total number of matching files

Details

Subject to matching on pattern finds the last modified file, and if verbose is TRUE reports on how many total files matched pattern.

Value

the name of the last modified file

Author(s)

Frank Harrell

See Also

base::list.files()

[Package Hmisc version 5.1-2 Index]