latexCheckOptions {Hmisc}	R Documentation

Check whether the options for latex functions have been specified.

Description

Check whether the options for latex functions have been specified. If any of
options()[c("latexcmd","dviExtension","xdvicmd")] are NULL, an error message is displayed.

Usage

latexCheckOptions(...)

Arguments

...
Any arguments are ignored.

Value

If any NULL options are detected, the invisible text of the error message. If all three options have non-NULL values, NULL.

Author(s)

Richard M. Heiberger <rmh@temple.edu>

See Also

latex

[Package Hmisc version 4.8-0 Index]