list.tree {Hmisc}	R Documentation

Pretty-print the Structure of a Data Object

Description

This is a function to pretty-print the structure of any data object (usually a list). It is similar to the R function str.

Usage

list.tree(struct, depth=-1, numbers=FALSE, maxlen=22, maxcomp=12, 
          attr.print=TRUE, front="", fill=". ", name.of, size=TRUE)

Arguments

struct
The object to be displayed

depth
Maximum depth of recursion (of lists within lists ...) to be printed; negative value means no limit on depth.

numbers
If TRUE, use numbers in leader instead of dots to represent position in structure.

maxlen
Approximate maximum length (in characters) allowed on each line to give the first few values of a vector. maxlen=0 suppresses printing any values.

maxcomp
Maximum number of components of any list that will be described.

attr.print
Logical flag, determining whether a description of attributes will be printed.

front
Front material of a line, for internal use.

fill
Fill character used for each level of indentation.

name.of
Name of object, for internal use (deparsed version of struct by default).

size
Logical flag, should the size of the object in bytes be printed?

A description of the structure of struct will be printed in outline form, with indentation for each level of recursion, showing the internal storage mode, length, class(es) if any, attributes, and first few elements of each data vector. By default each level of list recursion is indicated by a "." and attributes by "A".

Author(s)

Alan Zaslavsky, zaslavsk@hcp.med.harvard.edu

See Also

str

Examples

X <- list(a=ordered(c(1:30,30:1)),b=c("Rick","John","Allan"),
          c=diag(300),e=cbind(p=1008:1019,q=4))
list.tree(X)
# In R you can say str(X)

[Package Hmisc version 4.8-0 Index]