mtitle {Hmisc}	R Documentation

Margin Titles

Description

Writes overall titles and subtitles after a multiple image plot is drawn. If par()$oma==c(0,0,0,0), title is used instead of mtext, to draw titles or subtitles that are inside the plotting region for a single plot.

Usage

mtitle(main, ll, lc,  
       lr=format(Sys.time(),'%d%b%y'),
       cex.m=1.75, cex.l=.5, ...)

Arguments

main
main title to be centered over entire figure, default is none

ll
subtitle for lower left of figure, default is none

lc
subtitle for lower center of figure, default is none

lr
subtitle for lower right of figure, default is today's date in format 23Jan91 for UNIX or R (Thu May 30 09:08:13 1996 format for Windows). Set to "" to suppress lower right title.

cex.m
character size for main, default is 1.75

cex.l
character size for subtitles

...
other arguments passed to mtext

Value

nothing

Side Effects

plots

Author(s)

Frank Harrell
Department of Biostatistics, Vanderbilt University
fh@fharrell.com

See Also

par, mtext, title, unix, pstamp

Examples

#Set up for 1 plot on figure, give a main title,
#use date for lr
plot(runif(20),runif(20))
mtitle("Main Title")

#Set up for 2 x 2 matrix of plots with a lower left subtitle and overall title
par(mfrow=c(2,2), oma=c(3,0,3,0))
plot(runif(20),runif(20))
plot(rnorm(20),rnorm(20))
plot(exp(rnorm(20)),exp(rnorm(20)))
mtitle("Main Title",ll="n=20")

[Package Hmisc version 4.8-0 Index]