nCoincident {Hmisc}	R Documentation

nCoincident

Description

Number of Coincident Points

Usage

nCoincident(x, y, bins = 400)

Arguments

x
numeric vector

y
numeric vector

bins
number of bins in both directions

Details

Computes the number of x,y pairs that are likely to be obscured in a regular scatterplot, in the sense of overlapping pairs after binning into bins x bins squares where bins defaults to 400. NAs are removed first.

Value

integer count

Author(s)

Frank Harrell

Examples

nCoincident(c(1:5, 4:5), c(1:5, 4:5)/10)

[Package Hmisc version 5.1-2 Index]