num.intercepts {Hmisc}	R Documentation

Extract number of intercepts

Description

Extract the number of intercepts from a model

Usage

num.intercepts(fit, type=c('fit', 'var', 'coef'))

Arguments

fit
a model fit object

type
the default is to return the formal number of intercepts used when fitting the model. Set type='var' to return the actual number of intercepts stored in the var object, or type='coef' to return the actual number in the fitted coefficients. The former will be less than the number fitted for orm fits, and the latter for orm fits passed through fit.mult.impute

Value

num.intercepts returns an integer with the number of intercepts in the model.

See Also

orm, fit.mult.impute

[Package Hmisc version 4.8-0 Index]