print.char.list {Hmisc}	R Documentation

prints a list of lists in a visually readable format.

Description

Takes a list that is composed of other lists and matrixes and prints it in a visually readable format.

Usage

## S3 method for class 'char.list'
print(x, ..., hsep = c("|"), vsep = c("-"), csep = c("+"), print.it = TRUE,
                rowname.halign = c("left", "centre", "right"),
                rowname.valign = c("top", "centre", "bottom"),
                colname.halign = c("centre", "left", "right"),
                colname.valign = c("centre", "top", "bottom"),
                text.halign = c("right", "centre", "left"),
                text.valign = c("top", "centre", "bottom"),
                rowname.width, rowname.height,
                min.colwidth = .Options$digits, max.rowheight = NULL,
                abbreviate.dimnames = TRUE, page.width = .Options$width,
                colname.width, colname.height, prefix.width,
                superprefix.width = prefix.width)

Arguments

x
list object to be printed

...
place for extra arguments to reside.

hsep
character used to separate horizontal fields

vsep
character used to separate veritcal feilds

csep
character used where horizontal and veritcal separators meet.

print.it
should the value be printed to the console or returned as a string.

rowname.halign
horizontal justification of row names.

rowname.valign
verical justification of row names.

colname.halign
horizontal justification of column names.

colname.valign
verical justification of column names.

text.halign
horizontal justification of cell text.

text.valign
vertical justification of cell text.

rowname.width
minimum width of row name strings.

rowname.height
minimum height of row name strings.

min.colwidth
minimum column width.

max.rowheight
maximum row height.

abbreviate.dimnames
should the row and column names be abbreviated.

page.width
width of the page being printed on.

colname.width
minimum width of the column names.

colname.height
minimum height of the column names

prefix.width
maximum width of the rowname columns

superprefix.width
maximum width of the super rowname columns

Value

String that formated table of the list object.

Author(s)

Charles Dupont

[Package Hmisc version 4.8-0 Index]