print.princmp {Hmisc}	R Documentation

print.princmp

Description

Print Results of princmp

Usage

## S3 method for class 'princmp'
print(x, which = c("none", "standardized", "original", "both"), k = x$k, ...)

Arguments

x
results of princmp

which
specifies which loadings to print, the default being 'none' and other values being 'standardized', 'original', or 'both'

k
number of components to show, defaults to k specified to princmp

...
unused

Details

Simple print method for princmp()

Value

nothing

Author(s)

Frank Harrell

[Package Hmisc version 5.1-2 Index]