prnz {Hmisc}	R Documentation

Print and Object with its Name

Description

Prints an object with its name and with an optional descriptive text string. This is useful for annotating analysis output files and for debugging.

Usage

prn(x, txt, file)

Arguments

x
any object

txt
optional text string

file
optional file name. By default, writes to console. append=TRUE is assumed.

Side Effects

prints

See Also

print, cat

Examples

x <- 1:5
prn(x)
# prn(fit, 'Full Model Fit')

[Package Hmisc version 4.8-0 Index]