pstamp {Hmisc}	R Documentation

Date/Time/Directory Stamp the Current Plot

Description

Date-time stamp the current plot in the extreme lower right corner. Optionally add the current working directory and arbitrary other text to the stamp.

Usage

pstamp(txt, pwd = FALSE, time. = TRUE)

Arguments

txt
an optional single text string

pwd
set to TRUE to add the current working directory name to the stamp

time.
set to FALSE to use the date without the time

Details

Certain functions are not supported for S-Plus under Windows. For R, results may not be satisfactory if par(mfrow=) is in effect.

Author(s)

Frank Harrell

Examples

plot(1:20)
pstamp(pwd=TRUE, time=FALSE)

[Package Hmisc version 4.8-0 Index]