rMultinom {Hmisc}	R Documentation

Generate Multinomial Random Variables with Varying Probabilities

Description

Given a matrix of multinomial probabilities where rows correspond to observations and columns to categories (and each row sums to 1), generates a matrix with the same number of rows as has probs and with m columns. The columns represent multinomial cell numbers, and within a row the columns are all samples from the same multinomial distribution. The code is a modification of that in the impute.polyreg function in the MICE package.

Usage

rMultinom(probs, m)

Arguments

probs
matrix of probabilities

m
number of samples for each row of probs

Value

an integer matrix having m columns

See Also

rbinom

Examples

set.seed(1)
w <- rMultinom(rbind(c(.1,.2,.3,.4),c(.4,.3,.2,.1)),200)
t(apply(w, 1, table)/200)

[Package Hmisc version 4.8-0 Index]