seqFreq {Hmisc}	R Documentation

seqFreq

Description

Find Sequential Exclusions Due to NAs

Usage

seqFreq(..., labels = NULL, noneNA = FALSE)

Arguments

...
any number of variables

labels
if specified variable labels will be used in place of variable names

noneNA
set to TRUE to not include 'none' as a level in the result

Details

Finds the variable with the highest number of NAs. From the non-NAs on that variable find the next variable from those remaining with the highest number of NAs. Proceed in like fashion. The resulting variable summarizes sequential exclusions in a hierarchical fashion. See this for more information.

Value

factor variable with obs.per.numcond attribute

Author(s)

Frank Harrell

[Package Hmisc version 5.1-2 Index]