show.pch {Hmisc}	R Documentation

Display Colors, Plotting Symbols, and Symbol Numeric Equivalents

Description

show.pch plots the definitions of the pch parameters. show.col plots definitions of integer-valued colors. character.table draws numeric equivalents of all latin characters; the character on line xy and column z of the table has numeric code "xyz", which you would surround in quotes and preceed by a backslash.

Usage

show.pch(object = par("font"))
show.col(object=NULL)
character.table(font=1)

Arguments

object
font for show.pch, ignored for show.col.

font
font

Author(s)

Pierre Joyet pierre.joyet@bluewin.ch, Frank Harrell

See Also

points, text

Examples

## Not run: 
show.pch()
show.col()
character.table()

## End(Not run)

[Package Hmisc version 4.8-0 Index]