simplifyDims {Hmisc}	R Documentation

List Simplification

Description

Takes a list where each element is a group of rows that have been spanned by a multirow row and combines it into one large matrix.

Usage

simplifyDims(x)

Arguments

x
list of spanned rows

Details

All rows must have the same number of columns. This is used to format the list for printing.

Value

a matrix that contains all of the spanned rows.

Author(s)

Charles Dupont

See Also

rbind

Examples

a <- list(a = matrix(1:25, ncol=5), b = matrix(1:10, ncol=5), c = 1:5)

simplifyDims(a)

[Package Hmisc version 4.8-0 Index]