solvet {Hmisc}	R Documentation

solve Function with tol argument

Description

A slightly modified version of solve that allows a tolerance argument for singularity (tol) which is passed to qr.

Usage

solvet(a, b, tol=1e-09)

Arguments

a
a square numeric matrix

b
a numeric vector or matrix

tol
tolerance for detecting linear dependencies in columns of a

See Also

solve

[Package Hmisc version 4.8-0 Index]