src {Hmisc}	R Documentation

Source a File from the Current Working Directory

Description

src concatenates ".s" to its argument, quotes the result, and sources in the file. It sets options(last.source) to this file name so that src() can be issued to re-source the file when it is edited.

Usage

src(x)

Arguments

x
an unquoted file name aside from ".s". This base file name must be a legal S name.

Side Effects

Sets system option last.source

Author(s)

Frank Harrell

See Also

source

Examples

## Not run: 
src(myfile)   # source("myfile.s")
src()         # re-source myfile.s

## End(Not run)

[Package Hmisc version 4.8-0 Index]