string.bounding.box {Hmisc}	R Documentation

Determine Dimensions of Strings

Description

This determines the number of rows and maximum number of columns of each string in a vector.

Usage

string.bounding.box(string, type = c("chars", "width"))

Arguments

string
vector of strings

type
character: whether to count characters or screen columns

Value

rows
vector containing the number of character rows in each string

columns
vector containing the maximum number of character columns in each string

Author(s)

Charles Dupont

See Also

nchar, stringDims

Examples

a <- c("this is a single line string", "This is a\nmulti-line string")
stringDims(a)

[Package Hmisc version 4.8-0 Index]