string.break.line {Hmisc}	R Documentation

Break a String into Many Lines at Newlines

Description

Takes a string and breaks it into seperate substrings where there are newline characters.

Usage

string.break.line(string)

Arguments

string
character vector to be separated into many lines.

Value

Returns a list that is the same length of as the string argument.

Each list element is a character vector.

Each character vectors elements are the split lines of the corresponding element in the string argument vector.

Author(s)

Charles Dupont

See Also

strsplit

Examples

a <- c('', 'this is a single line string',
       'This is a\nmulti-line string.')

b <- string.break.line(a)

[Package Hmisc version 4.8-0 Index]