symbol.freq {Hmisc}	R Documentation

Graphic Representation of a Frequency Table

Description

This function can be used to represent contingency tables graphically. Frequency counts are represented as the heights of "thermometers" by default; you can also specify symbol='circle' to the function. There is an option to include marginal frequencies, which are plotted on a halved scale so as to not overwhelm the plot. If you do not ask for marginal frequencies to be plotted using marginals=T, symbol.freq will ask you to click the mouse where a reference symbol is to be drawn to assist in reading the scale of the frequencies.

label attributes, if present, are used for x- and y-axis labels. Otherwise, names of calling arguments are used.

Usage

symbol.freq(x, y, symbol = c("thermometer", "circle"),
            marginals = FALSE, orig.scale = FALSE,
            inches = 0.25, width = 0.15, subset, srtx = 0, ...)

Arguments

x
first variable to cross-classify

y
second variable

symbol
specify "thermometer" (the default) or "circle"

marginals
set to TRUE to add marginal frequencies (scaled by half) to the plot

orig.scale
set to TRUE when the first two arguments are numeric variables; this uses their original values for x and y coordinates)

inches
see symbols

width
see thermometers option in symbols

subset
the usual subsetting vector

srtx
rotation angle for x-axis labels

...
other arguments to pass to symbols

Author(s)

Frank Harrell

See Also

symbols

Examples

## Not run: 
getHdata(titanic)
attach(titanic)
age.tertile <- cut2(titanic$age, g=3)
symbol.freq(age.tertile, pclass, marginals=T, srtx=45)
detach(2)

## End(Not run)

[Package Hmisc version 4.8-0 Index]