sys {Hmisc}	R Documentation

Run Unix or Dos Depending on System

Description

Runs unix or dos depending on the current operating system. For R, just runs system with optional concatenation of first two arguments which are assumed named command and text.

Usage

sys(command, text=NULL, output=TRUE)
# S-Plus: sys(\dots, minimized=FALSE)

Arguments

command
system command to execute

text
text to concatenate to system command, if any (typically options or file names or both)

output
set to FALSE to not return output of command as a character vector

Value

see unix or dos

Side Effects

executes system commands

See Also

unix, system

[Package Hmisc version 4.8-0 Index]