upFirst {Hmisc}	R Documentation

Change First Letters to Upper Case

Description

Changes the first letter of each word in a string to upper case, keeping selected words in lower case. Words containing at least 2 capital letters are kept as-is.

Usage

upFirst(txt, lower = FALSE, alllower = FALSE)

Arguments

txt
a character vector

lower
set to TRUE to make only the very first letter of the string upper case, and to keep words with at least 2 capital letters in their original form

alllower
set to TRUE to make every word start with lower case unless it has at least 2 caps

References

https://en.wikipedia.org/wiki/Letter_case#Headings_and_publication_titles

Examples

upFirst(c('this and that','that is Beyond question'))

[Package Hmisc version 4.8-0 Index]