vlab {Hmisc}	R Documentation

vlab

Description

Easily Retrieve Text Form of Labels/Units

Usage

vlab(x, name = NULL)

Arguments

x
a single variable name, unquoted

name
optional character string to use as variable name

Details

Uses the same search method as hlab returns label and units in a character string with units, if present, in brackets

Value

character string

Author(s)

Frank Harrell

See Also

hlab()

[Package Hmisc version 5.1-2 Index]