xtfrm.labelled {Hmisc}	R Documentation

Auxiliary Function Method for Sorting and Ranking

Description

An auxiliary function method that is a workaround for bug in the implementation of xtfrm handles inheritance.

Usage

## S3 method for class 'labelled'
xtfrm(x)

Arguments

x
any object of class labelled.

See Also

xtfrm

[Package Hmisc version 4.8-0 Index]