yearDays {Hmisc}	R Documentation

Get Number of Days in Year or Month

Description

Returns the number of days in a specific year or month.

Usage

yearDays(time)

monthDays(time)

Arguments

time
A POSIXt or Date object describing the month or year in question.

Author(s)

Charles Dupont

See Also

POSIXt, Date

[Package Hmisc version 4.8-0 Index]