all.is.numeric {Hmisc}	R Documentation

Check if All Elements in Character Vector are Numeric

Description

Tests, without issuing warnings, whether all elements of a character vector are legal numeric values, or optionally converts the vector to a numeric vector. Leading and trailing blanks in x are ignored.

Usage

all.is.numeric(x, what = c("test", "vector"), extras=c('.','NA'))

Arguments

x
a character vector

what
specify what="vector" to return a numeric vector if it passes the test, or the original character vector otherwise

extras
a vector of character strings to count as numeric values, other than "".

Value

a logical value if what="test" or a vector otherwise

Author(s)

Frank Harrell

See Also

as.numeric

Examples

all.is.numeric(c('1','1.2','3'))
all.is.numeric(c('1','1.2','3a'))
all.is.numeric(c('1','1.2','3'),'vector')
all.is.numeric(c('1','1.2','3a'),'vector')
all.is.numeric(c('1','',' .'),'vector')

[Package Hmisc version 3.17-4 Index]