latexTabular {Hmisc}	R Documentation

Convert a Data Frame or Matrix to a LaTeX Tabular

Description

latexTabular creates a character vector representing a matrix or data frame in a simple tabular environment.

Usage

latexTabular(x, headings=colnames(x),
             align =paste(rep('c',ncol(x)),collapse=''),
             halign=paste(rep('c',ncol(x)),collapse=''),
             helvetica=TRUE, translate=TRUE, hline=0, center=FALSE, ...)

Arguments

x
a matrix or data frame, or a vector that is automatically converted to a matrix

headings
a vector of character strings specifying column headings for latexTabular, defaulting to x's colnames. To make multi-line headers use the newline character inside elements of headings.

align
a character strings specifying column alignments for latexTabular, defaulting to paste(rep('c',ncol(x)),collapse='') to center. You may specify align='c|c' and other LaTeX tabular formatting.

halign
a character strings specifying alignment for column headings, defaulting to centered.

helvetica
set to FALSE to use default LaTeX font in latexTabular instead of helvetica.

translate
set to FALSE if column headings and table entries are already in LaTeX format, otherwise latexTabular will run them through latexTranslate

hline
set to 1 to put hline after heading, 2 to also put hlines before and after heading and at table end

center
set to TRUE to enclose the tabular in a LaTeX center environment

...
if present, x is run through format.df with those extra arguments

Value

a character string containing LaTeX markup

Author(s)

Frank E. Harrell, Jr.,
Department of Biostatistics,
Vanderbilt University,
fh@fharrell.com

See Also

latex.default, format.df

Examples

x <- matrix(1:6, nrow=2, dimnames=list(c('a','b'),c('c','d','this that')))
latexTabular(x)   # a character string with LaTeX markup

[Package Hmisc version 4.8-0 Index]