makeNstr {Hmisc}	R Documentation

creates a string that is a repeat of a substring

Description

Takes a character and creates a string that is the character repeated len times.

Usage

makeNstr(char, len)

Arguments

char
character to be repeated

len
number of times to repeat char.

Value

A string that is char repeated len times.

Author(s)

Charles Dupont

See Also

paste, rep

Examples

makeNstr(" ", 5)

[Package Hmisc version 4.8-0 Index]