multLines {Hmisc}	R Documentation

Plot Multiple Lines

Description

Plots multiple lines based on a vector x and a matrix y, draws thin vertical lines connecting limits represented by columns of y beyond the first. It is assumed that either (1) the second and third columns of y represent lower and upper confidence limits, or that (2) there is an even number of columns beyond the first and these represent ascending quantiles that are symmetrically arranged around 0.5.

Usage

multLines(x, y, pos = c('left', 'right'), col='gray',
          lwd=1, lty=1, lwd.vert = .85, lty.vert = 1,
          alpha = 0.4, grid = FALSE)

Arguments

x
a numeric vector

y
a numeric matrix with number of rows equal to the number of x elements

pos
when pos='left' the vertical lines are drawn, right to left, to the left of the point (x, y[,1). Otherwise lines are drawn left to right to the right of the point.

col
a color used to connect (x, y[,1]) pairs. The same color but with transparency given by the alpha argument is used to draw the vertical lines

lwd
line width for main lines

lty
line types for main lines

lwd.vert
line width for vertical lines

lty.vert
line type for vertical lines

alpha
transparency

grid
set to TRUE when using grid/lattice

Author(s)

Frank Harrell

Examples

x <- 1:4
y <- cbind(x, x-3, x-2, x-1, x+1, x+2, x+3)
plot(NA, NA, xlim=c(1,4), ylim=c(-2, 7))
multLines(x, y, col='blue')
multLines(x, y, col='red', pos='right')

[Package Hmisc version 3.17-4 Index]