na.delete {Hmisc}	R Documentation

Row-wise Deletion na.action

Description

Does row-wise deletion as na.omit, but adds frequency of missing values for each predictor to the "na.action" attribute of the returned model frame. Optionally stores further details if options(na.detail.response=TRUE).

Usage

na.delete(frame)

Arguments

frame
a model frame

Value

a model frame with rows deleted and the "na.action" attribute added.

Author(s)

Frank Harrell
Department of Biostatistics
Vanderbilt University
f.harrell@vanderbilt.edu

See Also

na.omit, na.keep, na.detail.response, model.frame.default, naresid, naprint

Examples

# options(na.action="na.delete")
# ols(y ~ x)

[Package Hmisc version 3.17-4 Index]