nstr {Hmisc}	R Documentation

Creates a string of arbitry length

Description

Creates a vector of strings which consists of the string segment given in each element of the string vector repeated times.

Usage

nstr(string, times)

Arguments

string
character: vector of string segments to be repeated. Will be recycled if argument times is longer.

times
integer: vector of number of times to repeat the corisponding segment. Will be recycled if argument string is longer.

Value

returns a character vector the same length as the longest of the two arguments.

Note

Will throw a warning if the length of the longer argment is not a even multiple of the shorter argument.

Author(s)

Charles Dupont

See Also

paste, rep

Examples

nstr(c("a"), c(0,3,4))

nstr(c("a", "b", "c"), c(1,2,3))

nstr(c("a", "b", "c"), 4)

[Package Hmisc version 4.8-0 Index]